% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.r
\name{bayes_binom_test}
\alias{bayes_binom_test}
\title{Bayesian two-sample test of proportions}
\usage{
bayes_binom_test(x, n, a1, b1, a2, b2, a = a1, b = b1, pi0 = 0.5,
  pi1 = 1 - pi0, c1 = 1, c2 = 1, c = c1/c2, rule = pi1/pi0 * c2/c1)
}
\arguments{
\item{x}{two-vector containing successful trials}

\item{n}{number of trials (the same for both samples)}

\item{a1}{alpha, the hyperparameter of the beta distribution of the first poisson rate}

\item{b1}{beta, the hyperparameter of the beta distribution of the first poisson rate}

\item{a2}{alpha, the hyperparameter of the beta distribution of the second poisson rate}

\item{b2}{beta, the hyperparameter of the beta distribution of the second poisson rate}

\item{a}{alpha, the hyperparameter of the beta distribution under the null}

\item{b}{beta, the hyperparameter of the beta distribution under the null}

\item{pi0}{the prior probability of the null hypothesis}

\item{pi1}{the prior probability of the alternative hypothesis}

\item{c1}{loss associated with type I error}

\item{c2}{loss associated with type I error}

\item{c}{relative loss constant (loss due to type II error divided by loss due to type I error)}

\item{rule}{the decision rule for the Bayes factor B.  by default, test rejects when B < pi1/pi0 * c2/c1}
}
\description{
Bayesian two-sample test of proportions
}
\examples{
\dontrun{

a1 <- 3/5; b1 <- 7/5
a2 <- 7/5; b2 <- 3/5
plot_beta(c(a1, a2), c(b1, b2))
bayes_binom_test(c(12, 16), 20, a1, b1, a2, b2) # reject
bayes_binom_test(c(13, 16), 20, a1, b1, a2, b2) # fail to reject

# compare to :
prop.test(c(9, 16), c(20, 20)) # almost exact reject at 5\% level
prop.test(c(10, 16), c(20, 20)) # fail to reject
prop.test(c(11, 16), c(20, 20)) # fail to reject


# but even slightly more informative priors can overpower results
# watch the proportions to see why
a1 <- 3; b1 <- 7
a2 <- 7; b2 <- 3
plot_beta(c(a1, a2), c(b1, b2))
bayes_binom_test(c(12, 16), 20, a1, b1, a2, b2) # reject
bayes_binom_test(c(13, 16), 20, a1, b1, a2, b2) # reject
bayes_binom_test(c(16, 16), 20, a1, b1, a2, b2) # reject (!)
bayes_binom_test(c(17, 16), 20, a1, b1, a2, b2) # reject (!)
bayes_binom_test(c(18, 16), 20, a1, b1, a2, b2) # fail to reject (!)


}
}
\seealso{
\code{\link{sample_power}}, \code{\link{sample_power_est}}, \code{\link{find_size}}, \code{\link{plot_binom_rule}}
}

