% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.r
\name{dbetabinom}
\alias{dbetabinom}
\title{Density of the beta-binomial distribution with parameters size, alpha, and beta}
\usage{
dbetabinom(x, size, alpha, beta, log = FALSE)
}
\arguments{
\item{x}{observation}

\item{size}{the beta-binomial size parameter}

\item{alpha}{the beta-binomial alpha parameter}

\item{beta}{the beta-binomial beta parameter}

\item{log}{return log value?}
}
\description{
Density of the beta-binomial distribution with parameters size, alpha, and beta
}
\examples{
\dontrun{

dbetabinom(2, 10, .5, .5)
dbetabinom(0:10, 10, .5, .5)
sum(dbetabinom(0:10, 10, .5, .5))

}
}

