% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSize.r
\name{findSize}
\alias{findSize}
\title{Determine a sample size}
\usage{
findSize(power, a1, b1, a2, b2, a = a1, b = b1, pi0 = 0.5, pi1 = 1
  - pi0, c = 1, family = c("binomial", "poisson"), method = c("ones",
  "halves"), quiet = TRUE)
}
\arguments{
\item{power}{desired power}

\item{a1}{alpha, the hyperparameter of the prior distribution of the first poisson rate}

\item{b1}{beta, the hyperparameter of the prior distribution of the first poisson rate}

\item{a2}{alpha, the hyperparameter of the prior distribution of the second poisson rate}

\item{b2}{beta, the hyperparameter of the prior distribution of the second poisson rate}

\item{a}{alpha, the hyperparameter of the prior distribution under the null}

\item{b}{beta, the hyperparameter of the prior distribution under the null}

\item{pi0}{the prior probability of the null hypothesis}

\item{pi1}{the prior probability of the alternative hypothesis}

\item{c}{relative loss constant (loss due to type II error divided by loss due to type I error)}

\item{family}{"binomial" or "poisson", depending on test}

\item{method}{"ones" or "halves"}

\item{quiet}{message the user with calls to \code{\link{samplePower}}}
}
\description{
Determine a sample size for a given power and hyperparameter configuration
}
\examples{


\dontrun{

	
a1 <- 2; b1 <- 6
a2 <- 3; b2 <- 3
plotGamma(c(a1, a2), c(b1, b2))
findSize(power = .8, a1, b1, a2, b2, family = "poisson")

a1 <- 5; b1 <- .2 # E = 1
a2 <- 30; b2 <- .167 # E = 5
findSize(power = .8, a1, b1, a2, b2, family = "poisson")

a1 <- 4; b1 <- 2
a2 <- 6; b2 <- 2
findSize(power = .5, a1, b1, a2, b2)
findSize(power = .6, a1, b1, a2, b2)
findSize(power = .7, a1, b1, a2, b2)
findSize(power = .8, a1, b1, a2, b2) # takes a while
findSize(power = .8, a1, b1, a2, b2, method = "halves")

findSize(power = .9, a1, b1, a2, b2, method = "halves")


}

}
\seealso{
\code{\link{samplePower}}
}
