% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_dist.r
\name{plot_beta}
\alias{plot_beta}
\title{Plot a beta density}
\usage{
plot_beta(alpha, beta, n = 251, ...)
}
\arguments{
\item{alpha}{the beta's alpha parameter}

\item{beta}{the beta's beta parameter}

\item{n}{number of points to plot}

\item{...}{...}
}
\value{
a ggplot object
}
\description{
Plot a beta density
}
\details{
plot_beta creates a basic ggplot which can be styled.
}
\examples{
\dontrun{

plot_beta(1, 1)
plot_beta(.5, .5)
plot_beta(2, 2)
plot_beta(2, 2:3)

# nonuniqueness of modes
plot_beta(4.5, 7.5) + ylim(0, 3)
plot_beta(1.12, 1.22) + ylim(0, 3)
plot_beta(c(4.5, 1.12), c(7.5, 1.22)) + ylim(0, 3)


plot_beta(.5, .5) + theme_bw()
plot_beta(.5, .5) + theme_classic()

theme_set(theme_bw())
plot_beta(.5, .5) + labs(x = "theta", y = "")
plot_beta(.5, .5) + ggtitle("my plot")
plot_beta(.5, .5) + labs(x = expression(theta[1]), y = "belief")
plot_beta(.5, .5) + labs(x = expression(paste(theta[1], "= 5")), y = "belief")
plot_beta(.5, .5) + labs(x = expression(paste(theta[1], "= ", 5)), y = "belief")

}
}
\author{
David Kahle \email{david.kahle@gmail.com}
}

