% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_dist.r
\name{plot_gamma}
\alias{plot_gamma}
\title{Plot a gamma density}
\usage{
plot_gamma(alpha, beta, n = 251, q = 0.999, ...)
}
\arguments{
\item{alpha}{the gamma's alpha parameter}

\item{beta}{the gamma's beta parameter}

\item{n}{number of points to plot}

\item{q}{quantile used for xmax}

\item{...}{...}
}
\value{
a ggplot object
}
\description{
Plot a gamma density
}
\details{
plot_gamma creates a basic ggplot which can be styled.
}
\examples{
\dontrun{

plot_gamma(1, 1)
plot_gamma(.5, .5)
plot_gamma(2, 2)
plot_gamma(1:5, 1)
plot_gamma(1, 1:5)

plot_gamma(.5, .5, q = .999)
plot_gamma(.5, .5, q = .99)
plot_gamma(.5, .5, q = .9)


plot_gamma(2, 2) + theme_bw()
plot_gamma(2, 2) + theme_classic()
plot_gamma(2, c(2, 4)) + theme_classic()

theme_set(theme_bw())
plot_gamma(2, 2) + labs(x = "theta", y = "")
plot_gamma(2, 2) + ggtitle("my plot")
plot_gamma(2, 2) + labs(x = expression(lambda[1]), y = "belief")

}
}
\author{
David Kahle \email{david.kahle@gmail.com}
}

