% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDist.r
\name{plotBeta}
\alias{plotBeta}
\title{Plot a beta density}
\usage{
plotBeta(alpha, beta, n = 251, ...)
}
\arguments{
\item{alpha}{the beta's alpha parameter}

\item{beta}{the beta's beta parameter}

\item{n}{number of points to plot}

\item{...}{...}
}
\value{
a ggplot object
}
\description{
Plot a beta density
}
\details{
plotBeta creates a basic ggplot which can be styled.
}
\examples{

\dontrun{

plotBeta(1, 1)
plotBeta(.5, .5)
plotBeta(2, 2)
plotBeta(2, 2:3)

# nonuniqueness of modes
plotBeta(4.5, 7.5) + ylim(0, 3)
plotBeta(1.12, 1.22) + ylim(0, 3)
plotBeta(c(4.5, 1.12), c(7.5, 1.22)) + ylim(0, 3)


plotBeta(.5, .5) + theme_bw()
plotBeta(.5, .5) + theme_classic()

theme_set(theme_bw())
plotBeta(.5, .5) + labs(x = "theta", y = "")
plotBeta(.5, .5) + ggtitle("my plot")
plotBeta(.5, .5) + labs(x = expression(theta[1]), y = "belief")
plotBeta(.5, .5) + labs(x = expression(paste(theta[1], "= 5")), y = "belief")
plotBeta(.5, .5) + labs(x = expression(paste(theta[1], "= ", 5)), y = "belief")

}

}
\author{
David Kahle \email{david.kahle@gmail.com}
}
