% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDist.r
\name{plotGamma}
\alias{plotGamma}
\title{Plot a gamma density}
\usage{
plotGamma(alpha, beta, n = 251, q = 0.999, ...)
}
\arguments{
\item{alpha}{the gamma's alpha parameter}

\item{beta}{the gamma's beta parameter}

\item{n}{number of points to plot}

\item{q}{quantile used for xmax}

\item{...}{...}
}
\value{
a ggplot object
}
\description{
Plot a gamma density
}
\details{
plotGamma creates a basic ggplot which can be styled.
}
\examples{

\dontrun{

plotGamma(1, 1)
plotGamma(.5, .5)
plotGamma(2, 2)
plotGamma(1:5, 1)
plotGamma(1, 1:5)

plotGamma(.5, .5, q = .999)
plotGamma(.5, .5, q = .99)
plotGamma(.5, .5, q = .9)


plotGamma(2, 2) + theme_bw()
plotGamma(2, 2) + theme_classic()
plotGamma(2, c(2, 4)) + theme_classic()

theme_set(theme_bw())
plotGamma(2, 2) + labs(x = "theta", y = "")
plotGamma(2, 2) + ggtitle("my plot")
plotGamma(2, 2) + labs(x = expression(lambda[1]), y = "belief")

}

}
\author{
David Kahle \email{david.kahle@gmail.com}
}
