% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/powerPlot.r
\name{powerPlot}
\alias{powerPlot}
\title{Make a power curve}
\usage{
powerPlot(t_range, a1, b1, a2, b2, a = a1, b = b1, pi0 = 0.5, pi1 = 1 -
  pi0, ylim = 0:1, method = "fastest")
}
\arguments{
\item{t_range}{vector specifying x-range}

\item{a1}{alpha, the hyperparameter of the gamma distribution of the first poisson rate}

\item{b1}{beta, the hyperparameter of the gamma distribution of the first poisson rate}

\item{a2}{alpha, the hyperparameter of the gamma distribution of the second poisson rate}

\item{b2}{beta, the hyperparameter of the gamma distribution of the second poisson rate}

\item{a}{alpha, the hyperparameter of the gamma distribution under the null}

\item{b}{beta, the hyperparameter of the gamma distribution under the null}

\item{pi0}{the prior probability of the null hypothesis}

\item{pi1}{the prior probability of the alternative hypothesis}

\item{ylim}{y limits}

\item{method}{method argument to pass to \code{\link{samplePowerEst}}}
}
\description{
Make a power curve
}
\examples{
\dontrun{

a1 <- 2; b1 <- 6
a2 <- 3; b2 <- 3
powerPlot(t = c(0,1000), a1, b1, a2, b2)
powerPlot(t = c(0,100), a1, b1, a2, b2)
powerPlot(t = 0:100, a1, b1, a2, b2, method = 'exact')
# this last plot shows that the approximation is not quite exact for small values
# but decent for larger ones

# styling, all ggplot2 styling works
powerPlot(t = c(0,1000), a1, b1, a2, b2) +
  theme_bw() + opts(title = 'My Power Plot')




a1 <- 4; b1 <- 2
a2 <- 6; b2 <- 2
powerPlot(t = c(0,1000), a1, b1, a2, b2)
powerPlot(t = c(0,25), a1, b1, a2, b2)
powerPlot(t = 0:25, a1, b1, a2, b2, method = 'exact')




a1 <- c(2,4); b1 <- c(6,2);
a2 <- c(3,6); b2 <- c(3,2);
powerPlot(t = c(0,1000), a1, b1, a2, b2)

library(ggplot2)
last_plot() + theme_bw()




 }
}
\seealso{
\code{\link{samplePowerEst}}
}

