% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.r
\name{print.btest}
\alias{print.btest}
\title{Print a Bayesian test}
\usage{
\method{print}{btest}(x, digits = 4L, ...)
}
\arguments{
\item{x}{an object of class btest}

\item{digits}{number of digits to show}

\item{...}{additional parameters}
}
\value{
invisible object
}
\description{
Print a Bayesian test
}
\examples{
\dontrun{
  bayesBinomTest(11:12, 20, 1, 1, 1, 1)
}
}

