% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleAlpha.r
\name{sampleAlphaBinomial}
\alias{sampleAlphaBinomial}
\title{Compute the probability of type I error of the binomial proportions test for a given sample size (n) with specified beta priors}
\usage{
sampleAlphaBinomial(n, a1, b1, a2, b2, a = a1, b = b1, pi0 = 0.5,
  pi1 = 1 - pi0, c = 1)
}
\arguments{
\item{n}{the sample size}

\item{a1}{alpha, the hyperparameter of the prior distribution}

\item{b1}{beta, the hyperparameter of the prior distribution}

\item{a2}{alpha, the hyperparameter of the prior distribution}

\item{b2}{beta, the hyperparameter of the prior distribution}

\item{a}{alpha, the hyperparameter of the beta distribution under the null}

\item{b}{beta, the hyperparameter of the beta distribution under the null}

\item{pi0}{the prior probability of the null hypothesis}

\item{pi1}{the prior probability of the alternative hypothesis}

\item{c}{relative loss constant (loss due to type II error divided by loss due to type I error)}
}
\description{
Compute the probability of type I error of the binomial proportions test for a given sample size (n) with specified beta priors
}
\examples{
\dontrun{

a <- 5; b <- 5
sampleAlpha(t = 20, a, b)


}
}
\seealso{
\code{\link{sampleAlpha}}, \code{\link{findSize}}
}
