% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sampleAlpha.r
\name{sampleAlphaPoisson}
\alias{sampleAlphaPoisson}
\title{Compute the probability of type I error of the Poisson rate test for a given sample size (t) with specified gamma priors}
\usage{
sampleAlphaPoisson(t, a1, b1, a2, b2, a = a1, b = b1, pi0 = 0.5, pi1 = 1
  - pi0, c = 1)
}
\arguments{
\item{t}{the sample size}

\item{a1}{alpha, the hyperparameter of the gamma distribution of the first poisson rate}

\item{b1}{beta, the hyperparameter of the gamma distribution of the first poisson rate}

\item{a2}{alpha, the hyperparameter of the gamma distribution of the second poisson rate}

\item{b2}{beta, the hyperparameter of the gamma distribution of the second poisson rate}

\item{a}{alpha, the hyperparameter of the gamma distribution under the null}

\item{b}{beta, the hyperparameter of the gamma distribution under the null}

\item{pi0}{the prior probability of the null hypothesis}

\item{pi1}{the prior probability of the alternative hypothesis}

\item{c}{relative loss constant (loss due to type II error divided by loss due to type I error)}
}
\description{
Compute the probability of type I error of the Poisson rate test for a given sample size (t) with specified gamma priors
}
\examples{
\dontrun{

1+1

}
}
\seealso{
\code{\link{sampleAlpha}}, \code{\link{findSize}}
}

