% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePower.r
\name{samplePowerBinomial}
\alias{samplePowerBinomial}
\title{Compute the power of the binomial proportions test for a given sample size (n) with specified beta priors}
\usage{
samplePowerBinomial(n, a1, b1, a2, b2, a = a1, b = b1, pi0 = 0.5,
  pi1 = 1 - pi0, c = 1)
}
\arguments{
\item{n}{the sample size}

\item{a1}{alpha, the hyperparameter of the beta distribution of the first poisson rate}

\item{b1}{beta, the hyperparameter of the beta distribution of the first poisson rate}

\item{a2}{alpha, the hyperparameter of the beta distribution of the second poisson rate}

\item{b2}{beta, the hyperparameter of the beta distribution of the second poisson rate}

\item{a}{alpha, the hyperparameter of the beta distribution under the null}

\item{b}{beta, the hyperparameter of the beta distribution under the null}

\item{pi0}{the prior probability of the null hypothesis}

\item{pi1}{the prior probability of the alternative hypothesis}

\item{c}{relative loss constant (loss due to type II error divided by loss due to type I error)}
}
\description{
Compute the power of the binomial proportions test for a given sample size (n) with specified beta priors
}
\examples{
\dontrun{

# see ?samplePower


}
}
\seealso{
\code{\link{samplePower}}, \code{\link{samplePowerEst}}, \code{\link{findSize}}
}
