% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/samplePower.r
\name{samplePowerPoisson}
\alias{samplePowerPoisson}
\title{Compute the power of the Poisson rate test for a given sample size (t) with specified gamma priors}
\usage{
samplePowerPoisson(t, a1, b1, a2, b2, a = a1, b = b1, pi0 = 0.5, pi1 = 1
  - pi0, c = 1)
}
\arguments{
\item{t}{the sample size}

\item{a1}{alpha, the hyperparameter of the gamma distribution of the first poisson rate}

\item{b1}{beta, the hyperparameter of the gamma distribution of the first poisson rate}

\item{a2}{alpha, the hyperparameter of the gamma distribution of the second poisson rate}

\item{b2}{beta, the hyperparameter of the gamma distribution of the second poisson rate}

\item{a}{alpha, the hyperparameter of the gamma distribution under the null}

\item{b}{beta, the hyperparameter of the gamma distribution under the null}

\item{pi0}{the prior probability of the null hypothesis}

\item{pi1}{the prior probability of the alternative hypothesis}

\item{c}{relative loss constant (loss due to type II error divided by loss due to type I error)}
}
\description{
Compute the power of the Poisson rate test for a given sample size (t) with specified gamma priors
}
\examples{
\dontrun{

a1 <- 2; b1 <- 6
a2 <- 3; b2 <- 3
plot_gamma(c(a1, a2), c(b1, b2))
samplePowerPoisson(t = 10, a1, b1, a2, b2)
samplePowerPoisson(t = 11, a1, b1, a2, b2)
samplePowerPoisson(t = 10:11, a1, b1, a2, b2)
samplePowerPoisson(t = 250, a1, b1, a2, b2) # 56s
samplePowerEst(t = c(250, 500, 1000), a1, b1, a2, b2) # 5s, off by .001

a1 <- 4; b1 <- 2
a2 <- 6; b2 <- 2
plot_gamma(c(a1, a2), c(b1, b2))
samplePowerPoisson(t = 10, a1, b1, a2, b2)


}
}
\seealso{
\code{\link{samplePower}}, \code{\link{samplePowerEst}}, \code{\link{findSize}}
}

